Task Three/requirements.txt
opencv-python==4.7.0.68
mediapipe==0.9.0.1
scikit-learn==1.2.0
tensorflow==1.2.2
Pillow
tk
numpy

